update notebooks and pipelines
